........S. [ZBXNEXT-1638] changed types of zbx_strloc_t fields to size_t
git-svn-id: svn://svn.zabbix.com/branches/dev@59510 f4e7fd02-d210-4e34-a1e0-d4093fb66cc2

+++ ZBXNEXT-1638/include/common.h
@@ -1161,9 +1161,9 @@ unsigned int	zbx_alarm_off(void);
 typedef struct
 {
 	/* left position */
-	int	l;
+	size_t	l;
 	/* right position */
-	int	r;
+	size_t	r;
 }
 zbx_strloc_t;
